#31 [chore] 컴포넌트 별 파일 생성
