+++ app/src/main/java/com/hous/hous_aos/ui/newrules/component/CategoryItem.kt
@@ -1,0 +1,2 @@
+package com.hous.hous_aos.ui.newrules.component
+
