+++ app/src/main/java/com/hous/hous_aos/ui/newrules/component/NewRuleToolBar.kt
@@ -1,0 +1,55 @@
+package com.hous.hous_aos.ui.newrules.component
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontStyle
+import com.hous.hous_aos.R
+import com.hous.hous_aos.ui.newrules.NewRulesUiState
+import com.hous.hous_aos.ui.newrules.State
+
+@Composable
+fun NewRulesToolbar(
+    uiState: MutableState<NewRulesUiState>,
+    checkBoxState: MutableState<State>
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth(),
+        horizontalArrangement = Arrangement.SpaceBetween
+    ) {
+        Image(
+            painter = painterResource(id = R.drawable.ic_back),
+            contentDescription = ""
+        )
+        Text(
+            text = stringResource(id = R.string.new_rules_title),
+            fontStyle = FontStyle(R.style.B1)
+        )
+        if (uiState.value.notificationState && checkBoxState.value != State.SELECT) {
+            Image(
+                painter = painterResource(id = R.drawable.ic_alarmon),
+                contentDescription = "",
+                modifier = Modifier.clickable {
+                    uiState.value = uiState.value.copy(notificationState = false)
+                }
+            )
+        } else {
+            Image(
+                painter = painterResource(id = R.drawable.ic_alarmoff),
+                contentDescription = "",
+                modifier = Modifier.clickable {
+                    uiState.value = uiState.value.copy(notificationState = true)
+                }
+            )
+        }
+    }
+}
